re-named interface files, added namespaces for filters and tables
